# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Instantly.ai clone API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /labels:
    get:
      operationId: listLabels
      tags: [labels]
      summary: List all labels
      responses:
        "200":
          description: List of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
    post:
      operationId: createLabel
      tags: [labels]
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelInput"
      responses:
        "201":
          description: Created label
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
  /labels/{id}:
    delete:
      operationId: deleteLabel
      tags: [labels]
      summary: Delete a label
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/labels:
    put:
      operationId: setLeadLabels
      tags: [leads]
      summary: Replace the set of labels on a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLeadLabelsInput"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all email templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create an email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateInput"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        fromName:
          type: ["string", "null"]
        replyTo:
          type: ["string", "null"]
        dailyLimit:
          type: ["integer", "null"]
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        leadsCount:
          type: integer
        sentCount:
          type: integer
        openCount:
          type: integer
        clickCount:
          type: integer
        replyCount:
          type: integer
        bounceCount:
          type: integer
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        fromName:
          type: string
        replyTo:
          type: string
        dailyLimit:
          type: integer
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        fromName:
          type: string
        replyTo:
          type: string
        dailyLimit:
          type: integer
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
...
    Lead:
      type: object
      required: [id, email, createdAt, labels]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, unsubscribed, bounced, replied]
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        createdAt:
          type: string
    Label:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
    LabelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
    SetLeadLabelsInput:
      type: object
      required: [labelIds]
      properties:
        labelIds:
          type: array
          items:
            type: integer
...
    EmailTemplate:
      type: object
      required: [id, name, subject, body, bodyType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        previewText:
          type: ["string", "null"]
        body:
          type: string
        bodyType:
          type: string
          enum: [text, html]
        createdAt:
          type: string
    CreateTemplateInput:
      type: object
      required: [name, subject, body]
      properties:
        name:
          type: string
        subject:
          type: string
        previewText:
          type: string
        body:
          type: string
        bodyType:
          type: string
          enum: [text, html]
    UpdateTemplateInput:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        previewText:
          type: string
        body:
          type: string
        bodyType:
          type: string
          enum: [text, html]

    # ── Analytics ─────────────────────────────────────────────────────────────
...